chore: migrate browser tests to bun test; delete browser-manager (tested copy not source), colocate cdp-client
